+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## 0.28.0 - 2024-02-01
+
+#### Enhancements
+- Substantially increased iteration queue size
+- Added methods `DBNQueue.enable` and `DBNQueue.disable` for controlling queue consumption
+- Added method `DBNQueue.is_enabled` to signal the queue can accept records
+- Added method `DBNQueue.is_full` to signal the queue has reached capacity
+- Added enabled checks to `DBNQueue.put` and `DBNQueue.put_nowait`
+
+#### Breaking changes
+- Iterating a `Live` client after the streaming session has started will now raise a `ValueError`. Calling `Live.start` is not necessary when iterating the `Live` client
+- Moved constant `databento.live.client.DEFAULT_QUEUE_SIZE` to `databento.live.session.DBN_QUEUE_CAPACITY`
+- Removed `maxsize` parameter from `DBNQueue` constructor. `DBNQueue` now subclasses `SimpleQueue` instead
+- Removed property `DBNQueue.enabled`, use `DBNQueue.is_enabled` instead
+- Removed method `DBNQueue.is_half_full`, use `DBNQueue.is_full` instead
+
+#### Bug fixes
+- Fixed an issue where DBN records could be dropped while iterating
+- Fixed an issue where async iteration would block the event loop
+
 ## 0.27.0 - 2024-01-23
 
 #### Enhancements